@@ -24,7 +24,7 @@
     "react-router-dom": "^5.3.0",
     "react-scripts": "^4.0.3",
     "react-use-localstorage": "^3.5.3",
-    "typescript": "^4.1.2",
+    "typescript": "^4.4.4",
     "web-vitals": "^1.0.1"
   },
   "scripts": {
